Fix compatibility with main branch
